factor out DefferedStack
