+++ deferred.ts
@@ -1,6 +1,3 @@
-// taken from: https://deno.land/x/std@v0.2.11/util/deferred.ts
-// Copyright 2018-2019 the Deno authors. All rights reserved. MIT license.
-
 export type Deferred<T = any, R = Error> = {
   promise: Promise<T>;
   resolve: (t?: T) => void;
@@ -10,9 +7,8 @@ export type Deferred<T = any, R = Error> = {
 
 /** Create deferred promise that can be resolved and rejected by outside */
 export function defer<T>(): Deferred<T> {
-  let handled = false;
-  let resolve;
-  let reject;
+  let handled = false, resolve, reject;
+
   const promise = new Promise<T>((res, rej) => {
     resolve = r => {
       handled = true;
@@ -23,21 +19,47 @@ export function defer<T>(): Deferred<T> {
       rej(r);
     };
   });
+
   return {
     promise,
     resolve,
     reject,
+
     get handled() {
       return handled;
     }
   };
 }
 
-export function isDeferred(x): x is Deferred {
-  return (
-    typeof x === "object" &&
-    x.promise instanceof Promise &&
-    typeof x["resolve"] === "function" &&
-    typeof x["reject"] === "function"
-  );
+
+export class DeferredStack<T> {
+  private _array: Array<T>;
+  private _queue: Array<Deferred>;
+
+  constructor(ls?: Iterable<T>) {
+    this._array = ls ? [...ls] : [];
+    this._queue = [];
+  }
+
+  async pop(): Promise<T> {
+    if (this._array.length > 0) {
+      return this._array.pop();
+    }
+    const d = defer();
+    this._queue.push(d);
+    await d.promise;
+    return this._array.pop();
+  }
+
+  push(value: T): void {
+    this._array.push(value);
+    if (this._queue.length > 0) {
+      const d = this._queue.shift();
+      d.resolve();
+    }
+  }
+
+  get size(): number {
+    return this._array.length;
+  }
 }
